This is an initial commit of several over the next few days. The plan is to create a "master" version off which we all branch and work, only recommitting our updates/changes/fixes when there is agreement.
This current commit cleans up parts of the code and generalises it for the three different simulations we currently support (Millennium, Bolshoi, GiggleZ).

Note: To make things as machine independent as possible please note the change in assumed directory structure:

(1) TREES … symlink to the tree files for each simulation in ./input/treefiles/
(2) RESULTS ... output in ./output/results/ (see the *.par scripts).

In other words, no more hard linking to directories that don't exist on most people's computers.

DETAILS:

     modified:   .gitignore
Now ignore a few additional unnecessary files.

     modified:   Makefile
Added a flag for the GiggleZ simulation, and compiler options for running on gSTAR.

     modified:   cmd_file
Cleaned up this file and added paths to the bolshoi and giggles input files.

     modified:   code/core_simulation.h
Some cleaning and updating of the different simulations here.

     modified:   code/model_misc.c
Bolshoi spins currently have a problem with the x component. Until this is sorted out we need to add a fix for that simulation only.

     deleted:    gmon.out
This should not be version controlled, hence deleted and added to .gitignore.

     new file:   input/bolshoi.par
     deleted:    input/bolshoi_2006.par
     deleted:    input/bolshoi_2012.par
The old versions of the bolshoi input file have been merged into a single "best" version. The exact parameters of this best version will be finalised in the near future.

     new file:   input/gigglez.par
Added for the GiggleZ simulation. The "best" version is TBD.

     new file:   input/millennium.par
     deleted:    input/millennium_2006.par
     deleted:    input/millennium_2012.par
The old versions of the millennium input file have been merged into a single "best" version. The exact parameters of this best version will be finalised in the near future.

     modified:   input/desired_outputsnaps.txt
Updated for the three different simulations SAGE currently supports.

     modified:   output/allresults.py
The example allresults script has been cleaned and generalised for the different simulations and the results are now output in a specific "h" cosmology (not h'less anymore).

     renamed:    green_mpi.qsub -> sage.qsub
Renamed (to make it not machine specific) and minor updates.