The previous updates to the cooling code resulted in some unnoticed changes to the model results. These have been fixed with this update. In particular:
cooling: Even though the math was correct with the previous commit, it resulted in a change to the cooling radius by a factor of 2. This was subsequently acounted for in the later code, but not before the rcool>Rvir condition, producing different behaviour than expected. To account for this, the cooling time has been changed back to the dynamical time (a la the 2006 model).

parameter files: The bolshoi and millennium parameter files have been updated to account for the changes. In particular (and on a positive note), i've been able to find parameter choices that are almost identical between the two simulations (except for some BH physics).