This update enhances the radio mode heating prescription.
A heating radius has been added, which is the radius at which
0.5*m_heat_dot*Vvir^2 equals 0.5*m_cool_dot*Vvir^2. The cooling mass
is then decreased in proportion to the heating radius, assuming the
mass interior to the heating radius doesn't cool (since m prop r for
an isothermal profile). The heating radius is never allowed to decrease
(once heated always heated).

In effect, however, this is quite similar to the previous prescription.
Mostly the change is descriptive.

Also, i've turned on the bondi-hoyle accretion model as default.