This commit deals with the indexes needed to rebuild the SFR and Z histories in TAO.
Two new properties have been added to each galaxy/halo:

- mergeType marks the kind of merger that occured for the current galaxy. 0=none; 1=minor merger; 2=major merger; 3=disk instability; 4=disrupt to ICS
- mergeIntoID markd the galaxy into which this merger occured.

When any kind of merger happens, evolve_galaxies finds the previous instance (in time) of the galaxy in the halo and copies the above two values in. This way TAO knows what to do with such a branch as it tree-walks back in time (e.g. for a minor merger, add the SFR and Z at each point in the branch's past to the main progenitor's SFR and Z bulge history). Galaxies that are merged are removed from being output at that timestep, hence we need to add the merger info to the timestep before the merger occured.

To incorporate this there's been adjustments to the internal and external structures (as noted in the previous commit) and changes to how SAGE knows when to evolve a galaxy or ignore it.

I've also added back the propagation of halonr through to evolve_galaxies from main.c. It's usefule to have when checking that progenitor lines are behaving correctly.