HDF5 reader now works fully.
The read in of the tree is done in an incredibly sloppy manner.  This
was done mainly because I wanted to get it done and ensured it works.
In future, this should be updated to include macros/other functions so
code doesn't need to be copy pasted.

The code has been checked by running with binary mode and HDF5 mode and
doing a diff of the outputs.  This resulted in no differences.  I'm
about to check back out master, run the base version and do a final
check to ensure nothing is different there.